same shared structs as in the build scripts
